+++ src/scanner.c
@@ -101,8 +101,13 @@ Token scanToken(Scanner *scanner) {
     return makeToken(scanner,
                      match(scanner, '=') ? TOKEN_BANG_EQUAL : TOKEN_BANG);
   case '=':
-    return makeToken(scanner,
-                     match(scanner, '=') ? TOKEN_EQUAL_EQUAL : TOKEN_EQUAL);
+    if (match(scanner, '=')) {
+      return makeToken(scanner, TOKEN_EQUAL_EQUAL);
+    }
+    if (match(scanner, '>')) {
+      return makeToken(scanner, TOKEN_FAT_ARROW);
+    }
+    return makeToken(scanner, TOKEN_EQUAL);
   case '<':
     return makeToken(scanner,
                      match(scanner, '=') ? TOKEN_LESS_EQUAL : TOKEN_LESS);
@@ -264,7 +269,15 @@ static TokenType identifierType(Scanner *scanner) {
     }
     break;
   case 't':
-    return checkKeyword(scanner, 1, 3, "rue", TOKEN_TRUE);
+    if (scanner->current - scanner->start > 1) {
+      switch (scanner->start[1]) {
+      case 'r':
+        return checkKeyword(scanner, 2, 2, "ue", TOKEN_TRUE);
+      case 'y':
+        return checkKeyword(scanner, 2, 2, "pe", TOKEN_TYPE);
+      }
+    }
+    break;
   case 'v':
     return checkKeyword(scanner, 1, 2, "ar", TOKEN_VAR);
   case 'w':